refactor: split durable runtime lifecycle
